apply format and fix some raw dict checks

+++ qendpoint-core/src/main/java/com/the_qa_company/qendpoint/core/dictionary/impl/section/IntDictionarySection.java
@@ -414,6 +414,6 @@ public class IntDictionarySection implements DictionarySectionPrivate {
 
 	@Override
 	public DictionarySectionType getSectionType() {
-		return DictionarySectionType.FLOAT;
+		return DictionarySectionType.INT;
 	}
 }
